Show and define a statusline
Now the statusline is shown in vim displaying information like the
current git-branch, the file being edited and misc information from
file.

Perhaps making a new plugin with its own repository and installing it
through `Plug` would be a good idea, but this plugin is still pretty
straightforward and is not worth it.